requirements.txt
requests
feedparser
groq
python-dotenv